refactor kafka code snippet select
