+++ src/Kafka/KafkaInstanceDrawer/components/index.ts
@@ -1,14 +1,1 @@
-export * from "./JavaConfigCodeSnippet";
-export * from "./JavaConsumerCodeSnippet";
-export * from "./JavaProducerCodeSnippet";
-export * from "./PythonConfigCodeSnippet";
-export * from "./PythonConsumerCodeSnippet";
-export * from "./PythonProducerCodeSnippet";
-export * from "./QuarkusConfigCodeSnippet";
-export * from "./QuarkusConsumerCodeSnippet";
-export * from "./QuarkusProducerCodeSnippet";
-export * from "./SpringBootConsumerConfig";
-export * from "./SpringBootConsumerExample";
-export * from "./SpringBootListener";
-export * from "./SpringBootProducerCodeSnippet";
-export * from "./SprintBootConfigCodeSnippet";
+export * from "./SampleCodeSnippet";
